IBX-8784: Added `AbstractExceptionVisitor` template
